Add files via upload
🌟 Excited to share my first-semester final project! 🌟

I created MakeMyResume, a resume generator that simplifies the process of creating professional resumes. By taking user inputs, the tool arranges the information into a polished, professional format.

This project has been an incredible learning experience, and I would love to hear your thoughts and feedback on it. 🚀

Your suggestions would mean a lot as I continue to grow and refine my skills in this journey. 😊